@@ -2,13 +2,13 @@ source 'https://rubygems.org'
 
 ruby '2.7.5'
 
-gem 'rails', '5.2.6'
+gem 'rails', '6.1.6.1'
 
-gem 'coffee-rails', '~> 4.2.2'
-gem 'uglifier', '~> 3.2.0'
+gem 'coffee-rails'
+gem 'uglifier'
 
 gem 'sass-rails', '~> 5.0.6'
-gem 'compass-rails', '~> 4.0.0'
+gem 'compass-rails', '~> 3.0.0'
 
 gem 'jquery-rails', '~> 4.3.1'
 
