@@ -14,7 +14,7 @@ module Admin2::Design
       if @allowed_lp && @service.release_landing_page_version
         link = ActionController::Base.helpers.link_to I18n.t('admin2.landing_page.check_it_out'),
                                                       landing_page_without_locale_path, target: :_blank, rel: :noopener
-        flash[:notice] = I18n.t('admin2.landing_page.latest_version_released', link: link).html_safe # rubocop:disable Rails/OutputSafety
+        flash[:notice] = I18n.t('admin2.landing_page.latest_version_released', link: link).html_safe
       else
         flash[:error] = I18n.t('admin2.landing_page.this_version_is_not_released')
       end
