@@ -21,14 +21,14 @@ class Admin::ListingsService
   end
 
   def approve
-    listing.update_columns(state: Listing::APPROVED, # rubocop:disable Rails/SkipsModelValidations
+    listing.update_columns(state: Listing::APPROVED,
                            approval_count: listing.approval_count + 1)
     self.class.send_listing_approved(listing.id)
     notify_followers
   end
 
   def reject
-    listing.update_column(:state, Listing::APPROVAL_REJECTED) # rubocop:disable Rails/SkipsModelValidations
+    listing.update_column(:state, Listing::APPROVAL_REJECTED)
     self.class.send_listing_rejected(listing.id)
   end
 
