@@ -4,12 +4,12 @@ Rails.application.configure do
   APP_CONFIG ||= ConfigLoader.load_app_config
 
   # Settings specified here will take precedence over those in config/environment.rb
-
+  config.consider_all_requests_local = true
   # In the development environment your application's code is reloaded on
   # every request.  This slows down response time but is perfect for development
   # since you don't have to restart the webserver when you make code changes.
   config.cache_classes = false
-
+  config.hosts << /(.*\.)?lvh.me/
   # To autoload MailPreview, uncomment this line
   # (this is a hack which is fixed properly in Rails 4)
   # config.action_view.cache_template_loading = false
@@ -60,6 +60,7 @@ Rails.application.configure do
   # Enable/disable caching. By default caching is disabled.
   if Rails.root.join('tmp/caching-dev.txt').exist?
     config.action_controller.perform_caching = true
+    config.action_controller.enable_fragment_cache_logging = true
     config.cache_store = :memory_store, { :namespace => "sharetribe-dev"}
     config.public_file_server.headers = {
       'Cache-Control' => "public, max-age=#{2.days.to_i}"
@@ -97,6 +98,12 @@ Rails.application.configure do
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
 
+  # Raise exceptions for disallowed deprecations.
+  config.active_support.disallowed_deprecation = :raise
+
+  # Tell Active Support which deprecation messages to disallow.
+  config.active_support.disallowed_deprecation_warnings = []
+
   # Raise an error on page load if there are pending migrations.
   config.active_record.migration_error = :page_load
 
@@ -108,9 +115,11 @@ Rails.application.configure do
   config.assets.quiet = true
 
   # Raises error for missing translations
-  config.action_view.raise_on_missing_translations = false
-
+  config.i18n.raise_on_missing_translations = false
   # Use an evented file watcher to asynchronously detect changes in source code,
   # routes, locales, etc. This feature depends on the listen gem.
   config.file_watcher = ActiveSupport::EventedFileUpdateChecker
+
+  # List of classes deemed safe to be deserialized from YAML.
+  config.active_record.yaml_column_permitted_classes = [Symbol, ActiveSupport::HashWithIndifferentAccess]
 end
