@@ -76,5 +76,5 @@ ReactOnRails.configure do |config|
   # Client js uses assets not digested by rails.
   # For any asset matching this regex, non-digested symlink will be created
   # To disable symlinks set this parameter to nil.
-  config.symlink_non_digested_assets_regex = /\.(png|jpg|jpeg|gif|tiff|woff|ttf|eot|svg)/
+  # config.symlink_non_digested_assets_regex = /\.(png|jpg|jpeg|gif|tiff|woff|ttf|eot|svg)/
 end
